+++ Client/src/gui_button.cpp
@@ -3,8 +3,8 @@
 
 namespace PMG {
 	void GuiButton::Render(Renderer* renderer) {
-		renderer->FillRect(m_pos.x, m_pos.y, m_size.x, m_size.y, m_color);
-		renderer->RenderText(m_pos.x, m_pos.y, m_size.x, m_size.y, m_text);
+		renderer->FillRect(static_cast<int>(m_pos.x), static_cast<int>(m_pos.y), static_cast<int>(m_size.x), static_cast<int>(m_size.y), m_color);
+		renderer->RenderText(static_cast<int>(m_pos.x), static_cast<int>(m_pos.y), static_cast<int>(m_size.x), static_cast<int>(m_size.y), m_text);
 	}
 
 	void GuiButton::MousePressed(int x, int y) {
